Removed seperate components for stat view page. Created single file for all math to be done which will be accessed when/where needed. Added more statistics for future implementation
