+++ src/App.js
@@ -1,6 +1,6 @@
 import "./css/App.css";
 import Header from "./components/Header.js";
-import StatsView from "./components/StatsView.js";
+import StatsView from "./components/StatComponents/StatsView.js";
 import KilometerInput from "./components/KilometerInput.js";
 
 function App() {
