--- src/components/StatComponents/AveragePerDay.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AveragePerDay = () => {
-  return <p>Average kilometers run per day</p>;
-};
-
-export default AveragePerDay;
