--- src/components/StatComponents/KmsPerDayNeeded.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const KmsPerDayNeeded = () => {
-  return <p>Kilometers per day needed to hit the target</p>;
-};
-
-export default KmsPerDayNeeded;
