--- src/components/StatComponents/ProgressBar.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-const ProgressBar = () => {
-  return (
-    <>
-      {/* A placeholder for where the progress bar will go */}
-      <div className="progress-placeholder"></div>
-    </>
-  );
-};
-
-export default ProgressBar;
